🐛 Fix : 제거한 utils/responsive 파일에 대한 import 에러 해결
